requirements.txt
selenium
requests
datetime
goto-statement